Refactor StepList definitions again to be more clear
