+++ src/types/StepList/Step.d.ts
@@ -1,28 +1,8 @@
 import * as React from 'react';
-
-// Step Object is used in <Step> and <SubStep>
-export interface StepObject {
-  id?: string;
-  href: string;
-  title?: string;
-  heading: string;
-  headingLevel?: '1' | '2' | '3' | '4' | '5';
-  description?: string;
-  linkText?: string;
-  completed?: boolean;
-  started?: boolean;
-  isNextStep?: boolean;
-  onClick?: (...args: any[]) => any;
-  component?: React.ReactElement<any> | ((...args: any[]) => any);
-}
-
-// Substeps can be defined via the `steps` prop
-interface SubSteps {
-  steps?: StepObject[];
-}
+import { StepObjectWithSubStep } from './StepList';
 
 export interface StepProps {
-  step: StepObject & SubSteps
+  step: StepObjectWithSubStep;
   onStepLinkClick?: (...args: any[]) => any;
   showSubSubSteps?: boolean;
   completedText: string;
