+++ src/types/StepList/SubStep.d.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { StepObject } from './Step';
+import { StepObject } from './StepList';
 
 export interface SubStepProps {
   step: StepObject;
